requirements.txt
# Required Packages
flask
flask_sqlalchemy
flask_bcrypt
flask_login
